LPS-53355 Display page references are deleted when non default asset publisher is removed from a display page

+++ portal-impl/src/com/liferay/portlet/assetpublisher/AssetPublisherPortletLayoutListener.java
@@ -16,10 +16,12 @@ package com.liferay.portlet.assetpublisher;
 
 import com.liferay.portal.kernel.portlet.PortletLayoutListener;
 import com.liferay.portal.kernel.portlet.PortletLayoutListenerException;
+import com.liferay.portal.kernel.util.StringPool;
 import com.liferay.portal.model.Layout;
 import com.liferay.portal.model.PortletPreferences;
 import com.liferay.portal.service.LayoutLocalServiceUtil;
 import com.liferay.portal.service.SubscriptionLocalServiceUtil;
+import com.liferay.portlet.asset.util.AssetUtil;
 import com.liferay.portlet.assetpublisher.util.AssetPublisherUtil;
 import com.liferay.portlet.journal.service.JournalArticleLocalServiceUtil;
 
@@ -44,8 +46,12 @@ public class AssetPublisherPortletLayoutListener
 		try {
 			Layout layout = LayoutLocalServiceUtil.getLayout(plid);
 
-			JournalArticleLocalServiceUtil.deleteLayoutArticleReferences(
-				layout.getGroupId(), layout.getUuid());
+			if (AssetUtil.isDefaultAssetPublisher(
+					layout, portletId, StringPool.BLANK)) {
+
+				JournalArticleLocalServiceUtil.deleteLayoutArticleReferences(
+					layout.getGroupId(), layout.getUuid());
+			}
 
 			SubscriptionLocalServiceUtil.deleteSubscriptions(
 				layout.getCompanyId(), PortletPreferences.class.getName(),
